Remove `team_id` from Relfile
Because TeamID can be detected from `provisioning_profile`.
