+++ go/src/relparser/relfile/const.go
@@ -1,0 +1,5 @@
+package relfile
+
+const (
+	PREFIX string = "REL_CONFIG"
+)
